Simple RPG Character Builder
This is a simple RPG Character Builder. Input is done through parameters of the function call. There is no external user input built into this, but I'd like to come back to it and try to implement something that feature. Creating a GUI for an executable version would be fun to do as well!