Basic Caesar Cipher
This a very simple Caesar Cipher. Created through guided workshop to help build an understanding of defining functions and use of parameters/arguments.

I would like to expand on this eventually to see how complicated I could make an encryption like this and work on a more secure kind of encryption for practical use.